# perf(matrices): speed up eigenvals for triangular matrices