--- a/sympy/matrices/graph.py
+++ b/sympy/matrices/graph.py
@@ -64,6 +64,10 @@ def _strongly_connected_components(M):
     if not M.is_square:
         raise NonSquareMatrixError
 
+    rep = getattr(M, '_rep', None)
+    if rep is not None:
+        return rep.scc()
+
     V = range(M.rows)
     E = sorted(M.todok().keys())
     return strongly_connected_components((V, E))