--- a/sympy/matrices/graph.py
+++ b/sympy/matrices/graph.py
@@ -64,6 +64,7 @@ def _strongly_connected_components(M):
     if not M.is_square:
         raise NonSquareMatrixError
 
+    # RepMatrix uses the more efficient DomainMatrix.scc() method
     rep = getattr(M, '_rep', None)
     if rep is not None:
         return rep.scc()