--- a/sympy/polys/matrices/ddm.py
+++ b/sympy/polys/matrices/ddm.py
@@ -360,7 +360,8 @@ class DDM(list):
         See also
         ========
 
-        DomainMatrix.scc
+        sympy.polys.matrices.domainmatrix.DomainMatrix.scc
+
         """
         return a.to_sdm().scc()