<!-- Your title above should be a short description of what
was changed. Do not include the issue number in the title. -->

#### References to other Issues or PRs
<!-- If this pull request fixes an issue, write "Fixes #NNNN" in that exact
format, e.g. "Fixes #1234" (see
https://tinyurl.com/auto-closing for more information). Also, please
write a comment on that issue linking back to this pull request once it is
open. -->

Corrects a small performance regression from #21474

#### Brief description of what is fixed or changed

Speed up eigenvals for triangular matrices e.g.:
```python
def make(n):
    def entry(i, j):
        if i == j:
            return i
        elif i > j:
            return j
        else:
            return 0
    return Matrix(n, n, entry)

In [3]: M = make(5)

In [4]: M
Out[4]: 
⎡0  0  0  0  0⎤
⎢             ⎥
⎢0  1  0  0  0⎥
⎢             ⎥
⎢0  1  2  0  0⎥
⎢             ⎥
⎢0  1  2  3  0⎥
⎢             ⎥
⎣0  1  2  3  4⎦

In [5]: %time M.eigenvals()
CPU times: user 810 µs, sys: 313 µs, total: 1.12 ms
Wall time: 1.42 ms
Out[5]: {0: 1, 1: 1, 2: 1, 3: 1, 4: 1}
```
This had slowed down on master since 1.8 as a result of https://github.com/sympy/sympy/pull/21474 which is a performance improvement in general but slows down the particular case of a perfectly triangular matrix that factors completely into 1x1 blocks. The slowdown is due to calling calling charpoly and roots for trivial 1x1 matrices. This  PR just uses the value from the matrix directly as the eigenvalue if it is 1x1 which mostly restores the previous performance.

#### Other comments

This routine could become a lot faster for this kind of matrix after #21626.

#### Release Notes

<!-- Write the release notes for this release below between the BEGIN and END
statements. The basic format is a bulleted list with the name of the subpackage
and the release note for this PR. For example:

* solvers
  * Added a new solver for logarithmic equations.

* functions
  * Fixed a bug with log of integers.

or if no release note(s) should be included use:

NO ENTRY

See https://github.com/sympy/sympy/wiki/Writing-Release-Notes for more
information on how to write release notes. The bot will check your release
notes automatically to see if they are formatted correctly. -->

<!-- BEGIN RELEASE NOTES -->
NO ENTRY
<!-- END RELEASE NOTES -->