@oscarbenjamin 2021-07-28
A problem with the approach here:
```python
In [1]: M = Matrix([[(x + y)/z, 0], [0, x/z + y/z]])

In [2]: M
Out[2]: 
⎡x + y       ⎤
⎢─────    0  ⎥
⎢  z         ⎥
⎢            ⎥
⎢       x   y⎥
⎢  0    ─ + ─⎥
⎣       z   z⎦

In [3]: M.eigenvals()
Out[3]: 
⎧x + y     x   y   ⎫
⎨─────: 1, ─ + ─: 1⎬
⎩  z       z   z   ⎭
```
On master that gives:
```python
In [2]: M.eigenvals()
Out[2]: 
⎧x + y   ⎫
⎨─────: 2⎬
⎩  z     ⎭
```
It's possible to fix that here by adding a call to `cancel` but then the speed benefit is lost.

Probably a better way is that after #21626 we could do this only when a proper domain is being used rather than for EXRAW.