@smichr 2021-07-29
> possible to fix that here by adding a call

are there cases that need `cancel` or would `factor_terms` (which was made for the obvious cases) be sufficient:
```python
>>> factor_terms(x/z+y/z)
(x + y)/z
```