@oscarbenjamin 2021-07-29
In SymPy 1.8 we have this:
```python
In [21]: M = Matrix([[(x**2 - y**2)/(x - y), 0], [0, x + y]])

In [22]: M
Out[22]: 
⎡ 2    2       ⎤
⎢x  - y        ⎥
⎢───────    0  ⎥
⎢ x - y        ⎥
⎢              ⎥
⎣   0     x + y⎦

In [23]: M.eigenvals()
Out[23]: 
⎧ 2    2             ⎫
⎪x  - y              ⎪
⎨───────: 1, x + y: 1⎬
⎪ x - y              ⎪
⎩                    ⎭

In [24]: roots(M.charpoly())
Out[24]: {x + y: 2}
```
However on master that is
```python
In [2]: M.eigenvals()
Out[2]: {x + y: 2}

In [3]: roots(M.charpoly())
Out[3]: {x + y: 2}
```
Getting the multiplicity right is important because e.g. with sympy 1.8 we have
```python
In [1]: M = Matrix([[(x**2 - y**2)/(x - y), 0], [0, x + y]])

In [2]: M.exp()
---------------------------------------------------------------------------
NonInvertibleMatrixError
```
whereas master gives
```python
In [1]: M = Matrix([[(x**2 - y**2)/(x - y), 0], [0, x + y]])

In [2]: M.exp()
Out[2]: 
⎡ x + y        ⎤
⎢ℯ         0   ⎥
⎢              ⎥
⎢         x + y⎥
⎣  0     ℯ     ⎦
```
Currently PR would restore the buggy sympy 1.8 behaviour.