@smichr 2021-07-29
Does the following work without slowing things down too much?
```python
def merge_multiple(ev):
    from sympy.solvers.solvers import denoms
    from sympy.utilities.iterables import ordered
    from sympy.core.function import _mexpand
    from collections import defaultdict
    rv = {}
    d = defaultdict(list)
    for i in ev:
        d[tuple(ordered(i.free_symbols))].append(i)
    for i in d:
        if len(d[i]) == 1:
            rv[d[i]] = ev[d[i]]
            continue
        den = Mul(*denoms(Tuple(*d[i])), evaluate=False)
        uniq = defaultdict(list)
        for i in d[i]:
            uniq[_mexpand(den*i)].append(i)
        for i in uniq:
            k = next(ordered(uniq[i]))
            rv[k] = sum(ev[i] for i in uniq[i])
    return rv

    
>>> merge_multiple({(1/x+1/y):1,(x + y)/x/y:1})
{1/y + 1/x: 2}
>>> merge_multiple(Matrix([[(x**2 - y**2)/(x - y), 0], [0, x + y]]).eigenvals())
{x + y: 2}
```