@oscarbenjamin 2021-07-29
> Does the following work without slowing things down too much?

I'm not sure but I decided just to only apply this optimisation when the domain of the matrix is ZZ or QQ since then all expressions are in a canonical form.