@asmeurer 2021-07-29
Do the domains not include rational functions or other complex domains that also have canonical forms? Maybe we should add a `has_canonical` attribute to the domains. Or add `is_canonical` to the elements. That also opens the interesting possibility of not necessarily keeping things like rational functions in canonical form, which could be useful since cancelling after every operation might be slower than cancelling once at the end in some cases.