@oscarbenjamin 2021-07-29
> Do the domains not include rational functions or other complex domains that also have canonical forms?

They do but the implementation I just merged in #21626 only uses the domains automatically for ZZ and QQ. Otherwise the EXRAW domain is used precisely because the canonical forms would not be backwards compatible e.g. right now you can do:
```python
In [15]: M = Matrix([[(x + 1)**10000000000000]])

In [16]: M
Out[16]: 
⎡       10000000000000⎤
⎣(x + 1)              ⎦

In [17]: M**2
Out[17]: 
⎡       20000000000000⎤
⎣(x + 1)              ⎦
```
If that automatically used `ZZ[x]` then those binomials would need to be expanded.

Of course for charpoly/eigenvals we probably do want to use the domains but we need some code to identify the appropriate domain at the start of the routine because Matrix will not do it at construction.