@sylee957 2021-07-31
> You also need to change `_eigenvals_list`

I don’t think that this comment is resolved yet. I expect simultaneous change for eigenvals_list