@oscarbenjamin 2021-07-31
> I expect simultaneous change for eigenvals_list

I've made those changes now. Those functions should be merged somehow (one should call the other).